commit-modifiers/requirements.txt
--index https://pypi.tuna.tsinghua.edu.cn/simple
loguru==0.3.2
gitpython==2.1.13
tzlocal==2.0.0